// packages/datadog-api-client-v1/models/WidgetCustomLink.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
* Custom links help you connect a data value to a URL, like a Datadog page or your AWS console.
*/
export declare class WidgetCustomLink {
    /**
    * The flag for toggling context menu link visibility.
    */
    'isHidden'?: boolean;
    /**
    * The label for the custom link URL. Keep the label short and descriptive. Use metrics and tags as variables.
    */
    'label'?: string;
    /**
    * The URL of the custom link. URL must include `http` or `https`. A relative URL must start with `/`.
    */
    'link'?: string;
    /**
    * The label ID that refers to a context menu link. Can be `logs`, `hosts`, `traces`, `profiles`, `processes`, `containers`, or `rum`.
    */
    'overrideLabel'?: string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): WidgetCustomLink;
    static serialize(data: WidgetCustomLink): {
        [key: string]: any;
    };
    constructor();
}